lib/ukfile: API: Change iovec length to size_t
Previously the ukfile API would take the size of a passed iovec as a
signed int, a design oversight copied over from the binary syscall API.
Negative iovec lengths make no sense and should not be exposed by our
internal API.
This change makes iovec lengths unsigned size_t for ukfiles.
The posix-socket socket ops API is similarly changed to use size_t.
External socket implementations will need updating.

+++ lib/posix-eventfd/eventfd.c
@@ -35,7 +35,7 @@ struct evfd_alloc {
 
 
 static ssize_t evfd_read(const struct uk_file *f,
-			 const struct iovec *iov, int iovcnt,
+			 const struct iovec *iov, size_t iovcnt,
 			 size_t off, long flags __unused)
 {
 	int semaphore;
@@ -75,7 +75,7 @@ static ssize_t evfd_read(const struct uk_file *f,
 
 
 static ssize_t evfd_write(const struct uk_file *f,
-			  const struct iovec *iov, int iovcnt,
+			  const struct iovec *iov, size_t iovcnt,
 			  size_t off, long flags __unused)
 {
 	uint64_t add;
