+++ lib/ukfile/include/uk/file.h
@@ -36,7 +36,7 @@ struct uk_file;
 
 /* I/O */
 typedef ssize_t (*uk_file_io_func)(const struct uk_file *f,
-				   const struct iovec *iov, int iovcnt,
+				   const struct iovec *iov, size_t iovcnt,
 				   size_t off, long flags);
 
 /* Info (stat-like & chXXX-like) */
@@ -219,7 +219,7 @@ struct uk_file {
 /* Operations inlines */
 static inline
 ssize_t uk_file_read(const struct uk_file *f,
-		     const struct iovec *iov, int iovcnt,
+		     const struct iovec *iov, size_t iovcnt,
 		     size_t off, long flags)
 {
 	return f->ops->read(f, iov, iovcnt, off, flags);
@@ -227,7 +227,7 @@ ssize_t uk_file_read(const struct uk_file *f,
 
 static inline
 ssize_t uk_file_write(const struct uk_file *f,
-		      const struct iovec *iov, int iovcnt,
+		      const struct iovec *iov, size_t iovcnt,
 		      size_t off, long flags)
 {
 	return f->ops->write(f, iov, iovcnt, off, flags);
